KUDU-798 (part 5) Correct safe time advancement
This patch fixes safe time advancement in general and
allows for safe time advancement in the absence of writes
in particular.

The core of the patch is to plug in the new TimeManager
class wherever needed. There is an instance of a TimeManager
per tablet it's main uses are:

- When replicating messages to other replicas a leader
  uses the TimeManager to assign timestamps and to obtain
  a safe time to send, even when there are no writes.

- When receiving messages from a leader consensus uses
  the TimeManager to update the clock and to unblock
  any waiters that might waiting for a particular
  timestamp to be safe.

- Before a snapshot scan proceeds to scan, it must
  first wait for the TimeManager to deem whatever
  timestamp it has safe. Then it proceeds to wait for
  the snapshot at timestamp to have all its transactions
  committed and, finally, proceeds with the scan.

Put together, these changes allow to make sure that
snapshot scans are repeatable in the large majority
of cases.

Note: There is a high likely hood that https://gerrit.cloudera.org/#/c/5375/
will be merged with this patch. However I'm currently running
some experiments with it pre/post this patch.

Change-Id: I8532fdb069c8bee7f3e08ffe74cab0273885cc8e

WIP KUDU-1127 Don't hang scanner threads waiting for safe time

This makes it so we don't hang scanner threads on snapshot reads
when the difference between safe time and a requested timestamp
is bigger than the difference between 'now' and the client's
deadline. This makes the server return a Status::ServiceUnavailable()
with a THROTTLED error back to the client so that it can retry.

This allowed to swap linked_list-test to finish with snapshot scans
in the present and have it run on dist-test for 500 loops without
errors.

WIP as this requires a small directed test.

Change-Id: Ic7cd0b0749e715c5d9e665a8e37d0f1c95af574e